Get rust-la to compile and run tests against rust 1.0. No guarantees that anything actually works; and will need to do fixes to the api.

+++ src/internalutil.rs
@@ -1,9 +1,10 @@
-use std::num;
-use std::num::{Zero, One};
 use std::vec;
 
+use num;
+use num::{Float, Signed, One, Zero};
+
 #[inline]
-pub fn alloc_dirty_vec<T>(n : uint) -> Vec<T> {
+pub fn alloc_dirty_vec<T>(n : usize) -> Vec<T> {
   let mut v : Vec<T> = vec::Vec::with_capacity(n);
   unsafe {
     v.set_len(n);
